delete duplicate get_connector_files method
